@@ -1,10 +1,12 @@
-package com.pavel.a692group;
+package com.pavel.a692group.presentation.login;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.pavel.a692group.R;
 
 /**
  * Created by Pavel Mazhnik on 02.03.19.
